chore(components/ui/Table): move component file outside of stories folder
