+++ src/components/ui/__stories__/Table/Table.stories.tsx
@@ -2,12 +2,13 @@ import * as React from "react"
 import { Flex } from "@chakra-ui/react"
 import { Meta, StoryObj } from "@storybook/react"
 
+import { Table as TableComponent } from "../../Table"
+
 import {
   MdxDemoData,
   MdxEnergyConsumpData,
   MdxTypesOfBridgesData,
 } from "./mockMdxData"
-import { Table as TableComponent } from "./Table"
 
 const meta = {
   title: "Molecules / Display Content / ShadCN Tables",
